change allocation capacity to float (#67)
* change allocation caapcity to float

* Bump go version

* [auto generated] Checking in generated offline HTML doc

---------

+++ v1/pkg/client/model_service_storage_info.go
@@ -17,5 +17,5 @@ type ServiceStorageInfo struct {
 	// Allocated number of volumes.
 	AllocatedCount int32 `json:"AllocatedCount,omitempty"`
 	// Total capacity of all allocated volumes in TB for a service type.
-	AllocatedCapacity int32 `json:"AllocatedCapacity,omitempty"`
+	AllocatedCapacity float32 `json:"AllocatedCapacity,omitempty"`
 }
